+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Fixed
+
+- Super tic tac toe: fix timer tick overriding state
+
+### Changed
+
+- Super tic tac toe: better grid lines and clarity
+
 ## [0.2.7] - 2025-02-28
 
 ### Fixed